+++ R/PurpleAirQC_hourly_AB_00.R
@@ -1,14 +1,13 @@
 #' @export
 #' @importFrom rlang .data
 #' 
-#' @title Apply QC to Aggregation Statistics
+#' @title Apply hourly aggregation QC using "AB_OO" algorithm
 #' 
-#' @param aggregationStats Dataframe of statistics as returned by
-#' \code{pat_aggregate()}.
-#' @param min_count Aggregation bins with fewer than `min_count` measurements
-#' will be marked as `NA`.
+#' @param pat A PurpleAir timeseries object.
+#' @param min_count Aggregation bins with fewer than \code{min_count} measurements
+#' will be marked as \code{NA}.
 #' @param returnAllColumns Logical specifying whether to return all columns
-#' of statistical data generated for QC algorithm or just the final `pm25` 
+#' of statistical data generated for QC algorithm or just the final \code{pm25}
 #' result.
 #'  
 #' @description Creates a \code{pm25} timeseries by averaging aggregated data
@@ -29,6 +28,8 @@
 #' 
 #' @examples 
 #' \dontrun{
+#' library(AirSensor)
+#' 
 #' df <- 
 #'   example_pat %>%
 #'   pat_qc() %>%
@@ -46,22 +47,40 @@ PurpleAirQC_hourly_AB_00 <- function(
   # ----- Validate parameters --------------------------------------------------
   
   MazamaCoreUtils::stopIfNull(pat)
+  MazamaCoreUtils::stopIfNull(min_count)
+  MazamaCoreUtils::stopIfNull(returnAllColumns)
   
-  # ----- hourly_AB_01 ---------------------------------------------------------
+  # ----- Prepare aggregated data ----------------------------------------------
   
-  countData <- pat_aggregate(pat, function(x) { length(na.omit(x)) }) %>% 
+  # Hourly counts
+  countData <- 
+    pat %>% 
+    pat_aggregate(function(x) { base::length(na.omit(x)) }) %>% 
     pat_extractData()
-  meanData <- pat_aggregate(pat, function(x) { mean(x, na.rm = TRUE) }) %>% 
+  
+  # Hourly means
+  meanData <- 
+    pat %>%
+    pat_aggregate(function(x) { base::mean(x, na.rm = TRUE) }) %>% 
     pat_extractData()
+  
+  # ----- Create hourly dataframe ----------------------------------------------
+  
+  # NOTE:  Include variables used in QC so that they can be used to create a
+  # NOTE:  plot visualizing the how the QC algorithm rejects values.
   
   hourlyData <- 
     dplyr::tibble(datetime = meanData$datetime) %>%
+    
     # Create pm25 by averaging the A and B channel aggregation means
     dplyr::mutate(pm25 = (meanData$pm25_A + meanData$pm25_B) / 2) %>%
+    
     # Calculate min_count and mean_diff for use in QC
     dplyr::mutate(min_count = pmin(countData$pm25_A, countData$pm25_B, na.rm = TRUE)) %>%
     dplyr::mutate(mean_diff = abs(meanData$pm25_A - meanData$pm25_B)) %>%
-    # hourly_AB_00 follows
+    
+    # -----hourly_AB_00 QC algorithm -----
+    
     # When only a fraction of the data are reporting, something is wrong.
     # Invalidate data where:  (min_count < SOME_THRESHOLD)
     dplyr::mutate(pm25 = replace(
@@ -70,13 +89,13 @@ PurpleAirQC_hourly_AB_00 <- function(
       NA) 
     )
   
+  # ----- Return ---------------------------------------------------------------
+  
   if ( !returnAllColumns ) {
     hourlyData <- 
       hourlyData %>%
       dplyr::select(.data$datetime, .data$pm25)
   }
-  
-  # ----- Return ---------------------------------------------------------------
   
   return(hourlyData)